Update: Aspect-Ratio and Patterns Collection
- Previous commit resolved an issue where the `curses` writer's (x, y)
coordinates were opposite of my initial expectations, so the coordinates
were swapped (transposed).
- This updates the Game and Universe instances to support the new
aspect-ratio, as well as to support the new `patterns` module's class
to encapsulate Moore neighbourhood patterns with a radius of 1.
- Patterns are subject to the neighbourhood and game rules, so this is
an in-progress commit to do better to correlate the tacit assumptions of
the current design into explicit coordinations and parametrisations.